* commit '7a9102c7987603da415d6bd8185cc2d73bbc1ebd':
  Update manifest.yaml contents.
  Update documentation on installation steps and vNSF packaging
  Error messages on vNSF packages (non-targz / non-SHIELD)
  cURL scripts to quickly test Store API
  Docker improvements.
  Improved docker details & deployment.
  Included deployment validation.
  Presentation fix.
  Cleanup.
  Cleanup.
  Documentation update.
  Docker installation and documentation update.
  Docker installation and documentation update.
  Docker installation and documentation update.
  Docker compose to include MongoDB container.
  Work in progress
  Adding vNSF onboarding documentation.

+++ src/store.py
@@ -1,25 +1,11 @@
 # -*- coding: utf-8 -*-
 
-"""
-    Eve Demo
-    ~~~~~~~~
-
-    A demostration of a simple API powered by Eve REST API.
-
-    The live demo is available at eve-demo.herokuapp.com. Please keep in mind
-    that the it is running on Heroku's free tier using a free MongoHQ
-    sandbox, which means that the first request to the service will probably
-    be slow. The database gets a reset every now and then.
-
-    :copyright: (c) 2016 by Nicola Iarocci.
-    :license: BSD, see LICENSE for more details.
-"""
-
 import json
 
 import base64
 import os
 import shutil
+import store_errors as err
 import tarfile
 import tempfile
 import yaml
@@ -28,21 +14,18 @@ from flask import abort
 from werkzeug.datastructures import FileStorage, ImmutableMultiDict
 from werkzeug.utils import secure_filename
 
-# Heroku support: bind to PORT if defined, otherwise default to 5000.
-if 'PORT' in os.environ:
-    port = int(os.environ.get('PORT'))
-    # use '0.0.0.0' to ensure your REST API is reachable from all your
-    # network (and not only your computer).
-    host = '0.0.0.0'
-else:
-    port = 5000
-    host = '0.0.0.0'
+port = int(os.environ.get('API_PORT', 5000))
+
+# use '0.0.0.0' to ensure your REST API is reachable from all your
+# network (and not only your computer).
+host = '0.0.0.0'
 
 
 def onboard(request):
     # print(request.form)
     # print('files: {}'.format(request.files))
 
+    # TODO Define proper response codes according to errors.
     if 'package' not in request.files:
         abort(401)
 
@@ -54,33 +37,24 @@ def onboard(request):
     filename = secure_filename(file.filename)
     package_absolute_path = os.path.join(tempfile.gettempdir(), filename)
 
-    # print('I have {} with size {} type {}'.format(filename, file.content_length,
-    #                                               file.content_type))
-
     extracted_package_path = ''
 
     try:
-        #        file.save(package_absolute_path, buffer_size=1048576000)
         file.save(package_absolute_path)
 
         if not tarfile.is_tarfile(package_absolute_path):
-            # if not zipfile.is_zipfile(package_absolute_path):
-            file.remove()
-            abort(403)
-
-        # TODO remove this line (only here for large files try-out issues)
-        # shutil.rmtree(extracted_package_path)
+            abort(403, err.PKG_NOT_TARGZ)
 
         extracted_package_path = tempfile.mkdtemp()
-        #        os.makedirs(extracted_package_path)
 
-        # package = zipfile.ZipFile(package_absolute_path)
-        # package.extractall(extracted_package_path)
         package = tarfile.open(package_absolute_path)
         package.extractall(extracted_package_path)
         package.close()
 
-        stream = open(os.path.join(extracted_package_path, 'manifest.yaml'), 'r')
+        manifest_path = os.path.join(extracted_package_path, 'manifest.yaml')
+        if not os.path.isfile(manifest_path):
+            abort(403, err.PKG_NOT_SHIELD)
+        stream = open(manifest_path, 'r')
         manifest = dict(yaml.load(stream))
         stream.close()
 
@@ -92,9 +66,6 @@ def onboard(request):
         descriptor = stream.read()
         stream.close()
 
-        # TODO Store the proper file(s) and not the entire package
-        # The file will be closed automatically once stored.
-        #        xpto = open(package_absolute_path, 'rb')
         stream = open(os.path.join(extracted_package_path, 'manifest.yaml'), 'rb')
         fs = FileStorage(stream)
         files = request.files.copy()
@@ -104,24 +75,15 @@ def onboard(request):
         # Convert the vNSF package into the document data.
         # NOTE: there's no need to deep copy as the data won't be modified until it gets stored in the database.
         package_data = request.form.copy()
+        package_data['registry'] = {'vendor': "vNSF maker A", 'capabilities': ["some stuff"]}
         package_data['state'] = 'sandboxed'
         package_data['manifest'] = manifest
         package_data['descriptor'] = descriptor
         request.form = ImmutableMultiDict(package_data)
-
-
-        # xpto.close()
-        # # package = zipfile.ZipFile(os.path.join(extracted_package_path, 'x.zip'))
-        # # package.extractall(extracted_package_path)
-        # print(os.listdir(extracted_package_path))
-        # # xpto = open(os.path.join(extracted_package_path, 'x/x.zip'), 'r')
-        # # print('\n\n')
-        # # print(xpto)
-        # # print('\n\nDONE\n\n')
-
     finally:
         os.remove(package_absolute_path)
-        shutil.rmtree(extracted_package_path)
+        if os.path.isdir(extracted_package_path):
+            shutil.rmtree(extracted_package_path)
 
 
 def send_minimal_vnsf_data(response):
